# api-contract.yaml
openapi: 3.1.0
info:
  title: Routa.js API Contract
  description: |
    Single source of truth for the Routa.js dual-backend API.
    Both the Next.js backend (src/app/api/) and the Rust backend (crates/routa-server/)
    MUST implement all endpoints defined here with compatible request/response shapes.
  version: 0.1.0
servers:
  - url: http://localhost:3000
    description: Next.js backend (dev)
  - url: http://localhost:3210
    description: Rust backend (desktop)

# ─────────────────────────────────────────────────────────
# Shared component schemas
# ─────────────────────────────────────────────────────────
components:
  schemas:
...
    VirtualFileEntry:
      type: object
      required: [path, name, isDirectory]
      properties:
        path:
          type: string
        name:
          type: string
        isDirectory:
          type: boolean
        size:
          type: integer
        children:
          type: array
          items:
            $ref: "#/components/schemas/VirtualFileEntry"

    # ── Models ──
...
paths:
...
  /api/mcp:
    post:
      operationId: mcpPost
      summary: MCP Streamable HTTP (JSON-RPC)
      description: |
        Methods: initialize, tools/list, tools/call, notifications/initialized
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: JSON-RPC response
    get:
      operationId: mcpGet
      summary: MCP SSE stream
      responses:
        "200":
          description: SSE stream
          content:
            text/event-stream: {}
    delete:
      operationId: mcpDelete
      summary: Terminate MCP session
      responses:
        "200":
          description: Terminated
  /api/mcp/tools:
    get:
      operationId: listMcpTools
      summary: List MCP tools
      responses:
        "200":
          description: Tools list
    post:
      operationId: executeMcpTool
      summary: Execute an MCP tool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                args:
                  type: object
      responses:
        "200":
          description: Tool result
    patch:
      operationId: updateMcpToolsConfig
      summary: Update MCP tool configuration
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Updated config
...
  /api/clone:
    get:
      operationId: listClonedRepos
      summary: List cloned repositories
      responses:
        "200":
          description: Repo list
          content:
            application/json:
              schema:
                type: object
                properties:
                  repos:
                    type: array
                    items:
                      type: object
    post:
      operationId: cloneRepo
      summary: Clone a GitHub repository
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url]
              properties:
                url:
                  type: string
      responses:
        "200":
          description: Cloned
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  path:
                    type: string
                  name:
                    type: string
                  branch:
                    type: string
                  branches:
                    type: array
                    items:
                      type: string
                  existed:
                    type: boolean
    patch:
      operationId: switchBranch
      summary: Switch branch on cloned repo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [repoPath, branch]
              properties:
                repoPath:
                  type: string
                branch:
                  type: string
      responses:
        "200":
          description: Switched
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  branch:
                    type: string
                  branches:
                    type: array
                    items:
                      type: string
...
  /api/github/import:
    post:
      operationId: importGitHubRepo
      summary: Import a GitHub repo as a virtual workspace (zipball download)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                owner:
                  type: string
                repo:
                  type: string
                ref:
                  type: string
                  description: Branch, tag, or SHA. Defaults to HEAD.
                url:
                  type: string
                  description: Shorthand — GitHub URL (parsed into owner/repo)
      responses:
        "200":
          description: Import successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  owner:
                    type: string
                  repo:
                    type: string
                  ref:
                    type: string
                  fileCount:
                    type: integer
                  extractedPath:
                    type: string
                  importedAt:
                    type: string
                    format: date-time
        "404":
          description: Repository not found
        "403":
          description: Rate limited or forbidden
        "413":
          description: Repository too large
  /api/github/tree:
    get:
      operationId: getGitHubTree
      summary: Get file tree for an imported GitHub repo
      parameters:
        - name: owner
          in: query
          required: true
          schema:
            type: string
        - name: repo
          in: query
          required: true
          schema:
            type: string
        - name: ref
          in: query
          schema:
            type: string
            default: HEAD
      responses:
        "200":
          description: File tree
          content:
            application/json:
              schema:
                type: object
                properties:
                  tree:
                    type: array
                    items:
                      $ref: "#/components/schemas/VirtualFileEntry"
                  fileCount:
                    type: integer
        "404":
          description: Workspace not imported
  /api/github/file:
    get:
      operationId: getGitHubFile
      summary: Read a file from an imported GitHub repo
      parameters:
        - name: owner
          in: query
          required: true
          schema:
            type: string
        - name: repo
          in: query
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          schema:
            type: string
        - name: ref
          in: query
          schema:
            type: string
            default: HEAD
      responses:
        "200":
          description: File content
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string
                  path:
                    type: string
        "404":
          description: File or workspace not found
  /api/github/search:
    get:
      operationId: searchGitHubFiles
      summary: Search files in an imported GitHub repo
      parameters:
        - name: owner
          in: query
          required: true
          schema:
            type: string
        - name: repo
          in: query
          required: true
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
        - name: ref
          in: query
          schema:
            type: string
            default: HEAD
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  files:
                    type: array
                    items:
                      type: object
                      properties:
                        path:
                          type: string
                        name:
                          type: string
                        score:
                          type: number
                  total:
                    type: integer
                  query:
                    type: string

  # ── Workspace sub-resources ──
...
  /api/webhooks/github:
    get:
      operationId: listGithubWebhooks
      summary: List registered GitHub webhooks
      responses:
        "200":
          description: Webhook list
    post:
      operationId: handleGithubWebhook
      summary: Handle an incoming GitHub webhook event
      requestBody:
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Processed
...
  /api/github/pr-comment:
    post:
      operationId: postGithubPrComment
      summary: Post a comment on a GitHub pull request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [owner, repo, prNumber, body]
              properties:
                owner:
                  type: string
                repo:
                  type: string
                prNumber:
                  type: integer
                body:
                  type: string
      responses:
        "200":
          description: Comment posted

  # ── A2A Protocol ──
...